Remove CurlWebSocketTests (#5483)
